Added a load_stats_binary function to load stats proto from a file that stores the serialized proto.
Apparently tfdv.WriteStatisticsToText writes *serialized proto*, not text proto. And we don't have a corresponding load function.

Also changed the docstring of tfdv.load_statistics to reflect the fact.
Also renamed WriteStatsticsToText as WriteStatisticsToBinaryFile to reflect what it actually does. WriteStatisticsToText is now only an alias and will be removed in a future release.

PiperOrigin-RevId: 334248077