1) improved histogram accuracy
Previously, QUANTILES histograms were generated with identical counts per bucket. When the elements selected as quantiles boundaries are far from the idealized boundary - e.g., you have a distribution that's far from continuous - this was very wrong. This is a common enough case that it probably affects histogram accuracy for many users.

STANDARD histograms, being derived from the same quantiles source, were also affected.

This CL propagates the cumulative weight sum from the underlying quantiles sketch, and uses those values to fill in sample counts. Since what we get from the sketch is a sequence of bin upper bounds and associated weights, this requires changing Bucket semantics to include their upper bound and not lower bound, except for the first bin which includes both, because the quantiles sketch always gives us the minimum element and its count separately.

2) simplified infinity handling

The old handling of infinite values was complicated, and could result in bins that mixed finite and infinite values. For STANDARD histograms this is a problem, since we'd like to be able to align histograms using interpolation to calculate distance measures. I've updated STANDARD histogram generation to generate separate -inf and +inf bins, if applicable.

3) fixed nested list length custom stat

It looks like we were computing custom stats for nested list length based on the count of elements up a level in the nested list hierarchy, which I've fixed.

4) handles float64 overflow by omitting the standard histogram

PiperOrigin-RevId: 473787450